Project Clean up
