+++ Roguelite/Assets/Animation/Controllers/CharacterController.controller
@@ -162,7 +162,7 @@ AnimatorController:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: ThirdPersonAnimatorController
+  m_Name: CharacterController
   serializedVersion: 5
   m_AnimatorParameters:
   - m_Name: Forward
