+++ Roguelite/Assets/Animation/Controllers/CharacterController.controller.meta
