--- Roguelite/Assets/Scripts/Items/Shop Items.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 652c3b392cf45114eb56484055a56e51
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
