+++ Roguelite/Assets/Scripts/Movement/AICharacterControl.cs
@@ -4,11 +4,11 @@ using UnityEngine.AI;
 namespace Movement
 {
     [RequireComponent(typeof (NavMeshAgent))]
-    [RequireComponent(typeof (ThirdPersonCharacter))]
+    [RequireComponent(typeof (CharacterMovement))]
     internal sealed class AICharacterControl : MonoBehaviour
     {
         public NavMeshAgent Agent { get; private set; }
-        public ThirdPersonCharacter Character { get; private set; }
+        public CharacterMovement CharacterMovement { get; private set; }
 
         public Transform target;
 
@@ -16,7 +16,7 @@ namespace Movement
         private void Start()
         {
             Agent = GetComponentInChildren<NavMeshAgent>();
-            Character = GetComponent<ThirdPersonCharacter>();
+            CharacterMovement = GetComponent<CharacterMovement>();
 
 	        Agent.updateRotation = false;
 	        Agent.updatePosition = true;
@@ -27,7 +27,7 @@ namespace Movement
             if (target != null)
                 Agent.SetDestination(target.position);
 
-            Character.Move(Agent.remainingDistance > Agent.stoppingDistance ? Agent.desiredVelocity : Vector3.zero);
+            CharacterMovement.Move(Agent.remainingDistance > Agent.stoppingDistance ? Agent.desiredVelocity : Vector3.zero);
         }
 
         public void SetTarget(Transform target)
