--- Roguelite/Assets/Scripts/Shops.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 81c2a4bc3883ee84196811245c9fcb60
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
