--- Roguelite/Assets/Scripts/Test Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28c5b1c1b56e6ef4b80c5a767fb9b170
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
