--- Roguelite/Assets/Scripts/Test Scripts/HealthPot.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 56fd5e8b37d84e809f5b99bbdb93a56f
-timeCreated: 1583712325
