--- Roguelite/Assets/Scripts/Test Scripts/Sword.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 78a7e7b232ab80f42952cd10bdf7e0a0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
