--- Roguelite/Assets/Scripts/Test Scripts/SwordDef.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c61e4384f30bc8c4482cb04a04f30ae5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
