--- Roguelite/Assets/Scripts/Test Scripts/TestingScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f58972380dc08ab4d9798a958acdcd74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
